requirements.txt
anyio==4.9.0
asttokens==3.0.0
certifi==2025.4.26
charset-normalizer==3.4.2
colorama==0.4.6
executing==2.2.0
h11==0.16.0
httpcore==1.0.9
httpx==0.28.1
icecream==2.1.4
idna==3.10
nest-asyncio==1.6.0
Pygments==2.19.1
python-dotenv==1.1.0
python-telegram-bot==22.0
requests==2.32.3
sniffio==1.3.1
telegram==0.0.1
urllib3==2.4.0
